Projeto/pyproject.toml
[tool.poetry]
name = "projeto_how_bootcamp"
version = "0.1.0"
description = "Projeto bootcamp how"
authors = ["karinnecristina <karinnecristinapereira@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.3.4"
apache-airflow = "^2.2.2"
pytest = "^6.2.5"
lxml = "^4.7.1"
boto3 = "^1.20.24"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
